pyproject.toml
[project]
name = "test-python-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "curl-cffi>=0.10.0",
    "ipykernel>=6.29.5",
    "mysql-connector-python>=9.2.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "polars>=1.26.0",
    "psycopg2>=2.9.10",
    "pyarrow>=19.0.1",
    "requests>=2.32.3",
]
